Add observation of elapsed reads from the connection and received bytes metric
This commit adds observation of time spend on reading response from
the connection for variety of request types. The observation happens
from the header (first 8b) was read and attempt to read body was done.
It means that if during reading the body error was occuirred, metrics
will be reported. Please check sarama_broker_errors_total to see
if there were errors for that consumer.

This commit also adds metric of received bytes by each reaquest.